Implemented Core features and did some code refactoring

+++ lib/Data/Model/track.dart
@@ -6,7 +6,7 @@ class Track {
       largeImageUrl,
       date,
       externalUrl;
-  final double duration;
+  final int duration;
 
   Track(
       {required this.id,
